fixed detail screen voting
